refactor code to keep logic inside controllers, add users, profile, login, logout endpoints
